+++ fixtures/test-app/config.rb
@@ -1,4 +1,4 @@
 activate :google_analytics do |ga|
   ga.tracking_id = 'UA-123456-78'
-  ga.disable = true
+  ga.test = true
 end