+++ lib/middleman-google-analytics/helpers.rb
@@ -9,7 +9,7 @@ module Middleman
       def google_analytics_tag
         options = extensions[:google_analytics].options
 
-        return nil if legacy_development? && !options.development
+        return nil if options.disable
 
         file = File.join(File.dirname(__FILE__), 'analytics.js.erb')
         context = { options: options }
@@ -29,18 +29,6 @@ module Middleman
         str = "\n"
         content.each_line { |line| str << line.indent(2) }
         str
-      end
-
-      # Support for Middleman >= 3.4
-      def legacy_development?
-        # Middleman 3.4
-        is_development = try(:development?)
-        unless is_development.nil?
-          return is_development
-        end
-
-        # Middleman 4.x
-        app.development?
       end
 
     end